# public-executor/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for AttributeValueType.
const (
	Int    AttributeValueType = "int"
	String AttributeValueType = "string"
)

// Attribute defines model for Attribute.
type Attribute struct {
	Id        int64              `json:"id"`
	Name      string             `json:"name"`
	Slug      string             `json:"slug"`
	SortOrder int                `json:"sortOrder"`
	ValueType AttributeValueType `json:"valueType"`
}

// AttributeValueType defines model for Attribute.ValueType.
type AttributeValueType string

// AuthConfirmPasswordResetRequest defines model for AuthConfirmPasswordResetRequest.
type AuthConfirmPasswordResetRequest struct {
	NewPassword string `json:"newPassword"`
	Secret      string `json:"secret"`
}

// AuthLoginRequest defines model for AuthLoginRequest.
type AuthLoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`

	// Provider Идентификатор identity-провайдера. Пусто или local — email+password.
	Provider *string `json:"provider,omitempty"`
}

// AuthLoginResponse defines model for AuthLoginResponse.
type AuthLoginResponse struct {
	AccessExpiresAt time.Time `json:"accessExpiresAt"`
	AccessToken     string    `json:"accessToken"`

	// RefreshExpiresAt Сам refresh-секрет уходит в HttpOnly cookie; здесь — только срок действия для UI.
	RefreshExpiresAt time.Time `json:"refreshExpiresAt"`
	User             AuthUser  `json:"user"`
}

// AuthRegisterRequest defines model for AuthRegisterRequest.
type AuthRegisterRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// AuthRequestPasswordResetRequest defines model for AuthRequestPasswordResetRequest.
type AuthRequestPasswordResetRequest struct {
	Email string `json:"email"`
}

// AuthUser defines model for AuthUser.
type AuthUser struct {
	Email           string     `json:"email"`
	EmailVerifiedAt *time.Time `json:"emailVerifiedAt"`

	// Id public_id пользователя (UUID), он же sub в JWT.
	Id string `json:"id"`
}

// CartAddItemRequest defines model for CartAddItemRequest.
type CartAddItemRequest struct {
	ProductId int64 `json:"productId"`
	Quantity  int   `json:"quantity"`
}

// CartItem defines model for CartItem.
type CartItem struct {
	Product   *Product `json:"product,omitempty"`
	ProductId int64    `json:"productId"`
	Quantity  int      `json:"quantity"`
}

// CartRemoveItemRequest defines model for CartRemoveItemRequest.
type CartRemoveItemRequest struct {
	ProductId int64 `json:"productId"`
}

// CartResponse defines model for CartResponse.
type CartResponse struct {
	Items         []CartItem `json:"items"`
	TotalQuantity int        `json:"totalQuantity"`
}

// CartSetQuantityRequest defines model for CartSetQuantityRequest.
type CartSetQuantityRequest struct {
	ProductId int64 `json:"productId"`
	Quantity  int   `json:"quantity"`
}

// CategoryNode defines model for CategoryNode.
type CategoryNode struct {
	Children  []CategoryNode `json:"children"`
	Id        int64          `json:"id"`
	Name      string         `json:"name"`
	Slug      string         `json:"slug"`
	SortOrder int            `json:"sortOrder"`
}

// CategoryRef defines model for CategoryRef.
type CategoryRef struct {
	Id   int64  `json:"id"`
	Name string `json:"name"`
	Slug string `json:"slug"`
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	Items []OrderItemInput `json:"items"`
}

// ErrorDetailsModel defines model for ErrorDetailsModel.
type ErrorDetailsModel struct {
	Code    string `json:"code"`
	Details string `json:"details"`
	Message string `json:"message"`
}

// ErrorModel defines model for ErrorModel.
type ErrorModel struct {
	Error ErrorDetailsModel `json:"error"`
}

// GetCategoriesTreeResponse defines model for GetCategoriesTreeResponse.
type GetCategoriesTreeResponse struct {
	Items []CategoryNode `json:"items"`
}

// GetProductListResponse defines model for GetProductListResponse.
type GetProductListResponse struct {
	Items      []Product `json:"items"`
	TotalCount int       `json:"totalCount"`
}

// OrderItemInput defines model for OrderItemInput.
type OrderItemInput struct {
	ProductId int64 `json:"productId"`
	Quantity  int   `json:"quantity"`
}

// OrderItemResponse defines model for OrderItemResponse.
type OrderItemResponse struct {
	Name      string `json:"name"`
	ProductId int64  `json:"productId"`
	Quantity  int    `json:"quantity"`
}

// OrderResponse defines model for OrderResponse.
type OrderResponse struct {
	CreatedAt time.Time           `json:"createdAt"`
	Id        int64               `json:"id"`
	Items     []OrderItemResponse `json:"items"`
	Status    string              `json:"status"`
}

// Product defines model for Product.
type Product struct {
	Attributes []ProductAttributeValue `json:"attributes"`
	Category   *CategoryRef            `json:"category"`
	Id         int64                   `json:"id"`
	Images     []ProductImage          `json:"images"`
	Name       string                  `json:"name"`
	Slug       string                  `json:"slug"`
}

// ProductAttributeValue defines model for ProductAttributeValue.
type ProductAttributeValue struct {
	Attribute Attribute `json:"attribute"`
	ValueInt  *int      `json:"valueInt"`
	ValueText *string   `json:"valueText"`
}

// ProductImage defines model for ProductImage.
type ProductImage struct {
	IsPrimary bool   `json:"isPrimary"`
	SortOrder int    `json:"sortOrder"`
	Url       string `json:"url"`
}

// UploadProductImageResponse defines model for UploadProductImageResponse.
type UploadProductImageResponse struct {
	Id        int64  `json:"id"`
	IsPrimary bool   `json:"isPrimary"`
	Key       string `json:"key"`
	SortOrder int    `json:"sortOrder"`
	Url       string `json:"url"`
}

// GetProductsListParams defines parameters for GetProductsList.
type GetProductsListParams struct {
	// Category slug категории; вернётся пустой список, если категория неизвестна
	Category *string `form:"category,omitempty" json:"category,omitempty"`
}

// UploadProductImageMultipartBody defines parameters for UploadProductImage.
type UploadProductImageMultipartBody struct {
	File        openapi_types.File `json:"file"`
	IsPrimary   *bool              `json:"isPrimary,omitempty"`
	ProductSlug string             `json:"productSlug"`
	SortOrder   *int               `json:"sortOrder,omitempty"`
}

// AuthConfirmPasswordResetJSONRequestBody defines body for AuthConfirmPasswordReset for application/json ContentType.
type AuthConfirmPasswordResetJSONRequestBody = AuthConfirmPasswordResetRequest

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody = AuthLoginRequest

// AuthRegisterJSONRequestBody defines body for AuthRegister for application/json ContentType.
type AuthRegisterJSONRequestBody = AuthRegisterRequest

// AuthRequestPasswordResetJSONRequestBody defines body for AuthRequestPasswordReset for application/json ContentType.
type AuthRequestPasswordResetJSONRequestBody = AuthRequestPasswordResetRequest

// CartAddItemJSONRequestBody defines body for CartAddItem for application/json ContentType.
type CartAddItemJSONRequestBody = CartAddItemRequest

// CartRemoveItemJSONRequestBody defines body for CartRemoveItem for application/json ContentType.
type CartRemoveItemJSONRequestBody = CartRemoveItemRequest

// CartSetQuantityJSONRequestBody defines body for CartSetQuantity for application/json ContentType.
type CartSetQuantityJSONRequestBody = CartSetQuantityRequest

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// UploadProductImageMultipartRequestBody defines body for UploadProductImage for multipart/form-data ContentType.
type UploadProductImageMultipartRequestBody UploadProductImageMultipartBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /api/v1/public/auth/confirmPasswordReset)
	AuthConfirmPasswordReset(ctx echo.Context) error

	// (POST /api/v1/public/auth/login)
	AuthLogin(ctx echo.Context) error

	// (POST /api/v1/public/auth/logout)
	AuthLogout(ctx echo.Context) error

	// (POST /api/v1/public/auth/logoutAll)
	AuthLogoutAll(ctx echo.Context) error

	// (POST /api/v1/public/auth/refresh)
	AuthRefresh(ctx echo.Context) error

	// (POST /api/v1/public/auth/register)
	AuthRegister(ctx echo.Context) error

	// (POST /api/v1/public/auth/requestPasswordReset)
	AuthRequestPasswordReset(ctx echo.Context) error

	// (POST /api/v1/public/cart/addItem)
	CartAddItem(ctx echo.Context) error

	// (POST /api/v1/public/cart/clear)
	CartClear(ctx echo.Context) error

	// (GET /api/v1/public/cart/getCart)
	GetCart(ctx echo.Context) error

	// (POST /api/v1/public/cart/removeItem)
	CartRemoveItem(ctx echo.Context) error

	// (POST /api/v1/public/cart/setQuantity)
	CartSetQuantity(ctx echo.Context) error

	// (POST /api/v1/public/createOrder)
	CreateOrder(ctx echo.Context) error

	// (GET /api/v1/public/getCategoriesTree)
	GetCategoriesTree(ctx echo.Context) error

	// (GET /api/v1/public/getOrder/{id})
	GetOrder(ctx echo.Context, id int64) error

	// (GET /api/v1/public/getProduct/{slug})
	GetProduct(ctx echo.Context, slug string) error

	// (GET /api/v1/public/getProductsList)
	GetProductsList(ctx echo.Context, params GetProductsListParams) error

	// (POST /api/v1/public/uploadProductImage)
	UploadProductImage(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// AuthConfirmPasswordReset converts echo context to params.
func (w *ServerInterfaceWrapper) AuthConfirmPasswordReset(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthConfirmPasswordReset(ctx)
	return err
}

// AuthLogin converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogin(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogin(ctx)
	return err
}

// AuthLogout converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogout(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogout(ctx)
	return err
}

// AuthLogoutAll converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogoutAll(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogoutAll(ctx)
	return err
}

// AuthRefresh converts echo context to params.
func (w *ServerInterfaceWrapper) AuthRefresh(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthRefresh(ctx)
	return err
}

// AuthRegister converts echo context to params.
func (w *ServerInterfaceWrapper) AuthRegister(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthRegister(ctx)
	return err
}

// AuthRequestPasswordReset converts echo context to params.
func (w *ServerInterfaceWrapper) AuthRequestPasswordReset(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthRequestPasswordReset(ctx)
	return err
}

// CartAddItem converts echo context to params.
func (w *ServerInterfaceWrapper) CartAddItem(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CartAddItem(ctx)
	return err
}

// CartClear converts echo context to params.
func (w *ServerInterfaceWrapper) CartClear(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CartClear(ctx)
	return err
}

// GetCart converts echo context to params.
func (w *ServerInterfaceWrapper) GetCart(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetCart(ctx)
	return err
}

// CartRemoveItem converts echo context to params.
func (w *ServerInterfaceWrapper) CartRemoveItem(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CartRemoveItem(ctx)
	return err
}

// CartSetQuantity converts echo context to params.
func (w *ServerInterfaceWrapper) CartSetQuantity(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CartSetQuantity(ctx)
	return err
}

// CreateOrder converts echo context to params.
func (w *ServerInterfaceWrapper) CreateOrder(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateOrder(ctx)
	return err
}

// GetCategoriesTree converts echo context to params.
func (w *ServerInterfaceWrapper) GetCategoriesTree(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetCategoriesTree(ctx)
	return err
}

// GetOrder converts echo context to params.
func (w *ServerInterfaceWrapper) GetOrder(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int64

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetOrder(ctx, id)
	return err
}

// GetProduct converts echo context to params.
func (w *ServerInterfaceWrapper) GetProduct(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", ctx.Param("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter slug: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProduct(ctx, slug)
	return err
}

// GetProductsList converts echo context to params.
func (w *ServerInterfaceWrapper) GetProductsList(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetProductsListParams
	// ------------- Optional query parameter "category" -------------

	err = runtime.BindQueryParameter("form", true, false, "category", ctx.QueryParams(), &params.Category)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter category: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProductsList(ctx, params)
	return err
}

// UploadProductImage converts echo context to params.
func (w *ServerInterfaceWrapper) UploadProductImage(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UploadProductImage(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/v1/public/auth/confirmPasswordReset", wrapper.AuthConfirmPasswordReset)
	router.POST(baseURL+"/api/v1/public/auth/login", wrapper.AuthLogin)
	router.POST(baseURL+"/api/v1/public/auth/logout", wrapper.AuthLogout)
	router.POST(baseURL+"/api/v1/public/auth/logoutAll", wrapper.AuthLogoutAll)
	router.POST(baseURL+"/api/v1/public/auth/refresh", wrapper.AuthRefresh)
	router.POST(baseURL+"/api/v1/public/auth/register", wrapper.AuthRegister)
	router.POST(baseURL+"/api/v1/public/auth/requestPasswordReset", wrapper.AuthRequestPasswordReset)
	router.POST(baseURL+"/api/v1/public/cart/addItem", wrapper.CartAddItem)
	router.POST(baseURL+"/api/v1/public/cart/clear", wrapper.CartClear)
	router.GET(baseURL+"/api/v1/public/cart/getCart", wrapper.GetCart)
	router.POST(baseURL+"/api/v1/public/cart/removeItem", wrapper.CartRemoveItem)
	router.POST(baseURL+"/api/v1/public/cart/setQuantity", wrapper.CartSetQuantity)
	router.POST(baseURL+"/api/v1/public/createOrder", wrapper.CreateOrder)
	router.GET(baseURL+"/api/v1/public/getCategoriesTree", wrapper.GetCategoriesTree)
	router.GET(baseURL+"/api/v1/public/getOrder/:id", wrapper.GetOrder)
	router.GET(baseURL+"/api/v1/public/getProduct/:slug", wrapper.GetProduct)
	router.GET(baseURL+"/api/v1/public/getProductsList", wrapper.GetProductsList)
	router.POST(baseURL+"/api/v1/public/uploadProductImage", wrapper.UploadProductImage)

}