# public-executor/api/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Decoupage Napkins Shop API server
paths:
  /api/v1/public/createOrder:
    post:
      tags: [Orders]
      operationId: createOrder
      description: 'Оформить заказ'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: 'Заказ создан'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/getOrder/{id}:
    get:
      tags: [Orders]
      operationId: getOrder
      description: 'Получить заказ по ID'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/getProduct/{slug}:
    get:
      tags: [Products]
      operationId: getProduct
      description: 'Получить карточку товара по slug'
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/getProductsList:
    get:
      tags: [Products]
      operationId: getProductsList
      description: 'Список товаров (опционально с фильтром по slug категории)'
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: 'slug категории; вернётся пустой список, если категория неизвестна'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/getCategoriesTree:
    get:
      tags: [Categories]
      operationId: getCategoriesTree
      description: 'Двухуровневое дерево категорий: корни + их дочерние'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCategoriesTreeResponse'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/auth/register:
    post:
      tags: [Auth]
      operationId: authRegister
      description: 'Регистрация по email и паролю. Возвращает пользователя и пару токенов; refresh-секрет дополнительно ставится HttpOnly cookie.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '409':
          description: Conflict (email taken)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      description: 'Логин по email и паролю. Если в запросе присутствует guest-cart cookie, гостевая корзина мерджится в пользовательскую.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/auth/refresh:
    post:
      tags: [Auth]
      operationId: authRefresh
      description: 'Обмен refresh-секрета на новую пару токенов. Секрет читается из HttpOnly cookie; rotation: старый секрет помечается отозванным.'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
  /api/v1/public/auth/logoutAll:
    post:
      tags: [Auth]
      operationId: authLogoutAll
      description: 'Отзыв всех активных refresh-секретов пользователя (все устройства). Требует валидный access JWT.'
      responses:
        '204':
          description: 'OK'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/auth/requestPasswordReset:
    post:
      tags: [Auth]
      operationId: authRequestPasswordReset
      description: 'Создаёт одноразовый токен восстановления. Реакция всегда 204 (без раскрытия существования email).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestPasswordResetRequest'
      responses:
        '204':
          description: 'OK'
  /api/v1/public/auth/confirmPasswordReset:
    post:
      tags: [Auth]
      operationId: authConfirmPasswordReset
      description: 'Подтверждает восстановление пароля по секрету. Все активные refresh-токены пользователя отзываются.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfirmPasswordResetRequest'
      responses:
        '204':
          description: 'OK'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
  /api/v1/public/cart/addItem:
    post:
      tags: [Cart]
      operationId: cartAddItem
      description: 'Добавить товар в корзину (или увеличить количество).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartAddItemRequest'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /api/v1/public/cart/setQuantity:
    post:
      tags: [Cart]
      operationId: cartSetQuantity
      description: 'Установить точное количество товара (qty <= 0 удалит позицию).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartSetQuantityRequest'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /api/v1/public/cart/removeItem:
    post:
      tags: [Cart]
      operationId: cartRemoveItem
      description: 'Удалить позицию из корзины.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRemoveItemRequest'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
...
  /api/v1/public/uploadProductImage:
    post:
      tags: [Products]
      operationId: uploadProductImage
      description: 'Загрузить картинку товара в S3 и сохранить ссылку'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, productSlug]
              properties:
                file:
                  type: string
                  format: binary
                productSlug:
                  type: string
                isPrimary:
                  type: boolean
                sortOrder:
                  type: integer
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProductImageResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
components:
  schemas:
    GetProductListResponse:
      type: object
      required: [items, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        totalCount:
          type: integer
    Product:
      type: object
      required: [id, slug, name, attributes, images]
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        name:
          type: string
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/ProductAttributeValue'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        category:
          $ref: '#/components/schemas/CategoryRef'
    ProductImage:
      type: object
      required: [url, isPrimary, sortOrder]
      properties:
        url:
          type: string
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
    UploadProductImageResponse:
      type: object
      required: [id, key, url, isPrimary, sortOrder]
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        url:
          type: string
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
    CategoryRef:
      type: object
      nullable: true
      required: [id, slug, name]
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        name:
          type: string
...
    ProductAttributeValue:
      type: object
      required: [attribute]
      properties:
        attribute:
          $ref: '#/components/schemas/Attribute'
        valueText:
          type: string
          nullable: true
        valueInt:
          type: integer
          nullable: true
    GetCategoriesTreeResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryNode'
    CategoryNode:
      type: object
      required: [id, slug, name, sortOrder, children]
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        name:
          type: string
        sortOrder:
          type: integer
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategoryNode'
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemInput'
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    OrderResponse:
      type: object
      required: [id, items, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    OrderItemResponse:
      type: object
      required: [productId, name, quantity]
      properties:
        productId:
          type: integer
          format: int64
        name:
          type: string
        quantity:
          type: integer
    AuthRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        provider:
          type: string
          description: 'Идентификатор identity-провайдера. Пусто или local — email+password.'
        email:
          type: string
        password:
          type: string
    AuthLoginResponse:
      type: object
      required: [user, accessToken, accessExpiresAt, refreshExpiresAt]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        accessToken:
          type: string
        accessExpiresAt:
          type: string
          format: date-time
        refreshExpiresAt:
          type: string
          format: date-time
          description: 'Сам refresh-секрет уходит в HttpOnly cookie; здесь — только срок действия для UI.'
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          description: 'public_id пользователя (UUID), он же sub в JWT.'
        email:
          type: string
        emailVerifiedAt:
          type: string
          format: date-time
          nullable: true
    AuthRequestPasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthConfirmPasswordResetRequest:
      type: object
      required: [secret, newPassword]
      properties:
        secret:
          type: string
        newPassword:
          type: string
    CartResponse:
      type: object
      required: [items, totalQuantity]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        totalQuantity:
          type: integer
...
    CartAddItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    CartSetQuantityRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 0
    CartRemoveItemRequest:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
          format: int64
    ErrorModel:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetailsModel'
    ErrorDetailsModel:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string